refactoring homework-2
