+++ src/components/BookList/BookListItem.vue
