+++ src/views/general/PageTemplate/PageTemplate.vue
@@ -4,5 +4,3 @@
     <slot name="main" />
   </div>
 </template>
-
-<style lang="scss" scoped></style>
